refactor: move schedule creation to the schedule.py
